improve typing
