+++ src/hooks/useLikedPokemons.ts
@@ -1,9 +1,8 @@
-import { PokemonTypes } from "@/src/constants";
-
 import { useQuery } from "@apollo/client";
 import { useCallback, useState } from "react";
 import { gql } from "../__generated__";
 import { useLikeContext } from "../contexts/LikedContext";
+import { parsePokemon } from "./utils";
 
 const GET_POKEMON_INDEX = gql(`
   query liked_pokemons($offset: Int, $limit: Int, $ids: [Int!]) {
@@ -38,15 +37,7 @@ const useLikedPokemons = () => {
 
   const pokemonIndex =
     data?.pokemon_v2_pokemon.map((item) => {
-      return {
-        id: item.id,
-        name: item.name,
-        imageUri: item.pokemon_v2_pokemonsprites[0].sprites["official-artwork"]
-          .front_default as string,
-        types: item.pokemon_v2_pokemontypes
-          ?.map((item) => item.pokemon_v2_type?.name)
-          .filter((item) => item) as PokemonTypes[],
-      };
+      return parsePokemon(item);
     }) ?? [];
 
   const handleLoadMore = useCallback(async () => {
