+++ src/hooks/usePokemon.ts
@@ -1,8 +1,7 @@
-import { PokemonTypes } from "@/src/constants";
-
 import { useQuery } from "@apollo/client";
 import { gql } from "../__generated__";
 import { PokemonQuery } from "../__generated__/graphql";
+import { parsePokemon } from "./utils";
 
 const GET_POKEMON = gql(`
   query pokemon($name: String) {
@@ -29,25 +28,6 @@ const GET_POKEMON = gql(`
   }
 `);
 
-type Pokemon = PokemonQuery["pokemon_v2_pokemon"][number];
-
-const parsePokemon = (pokemon?: Pokemon) => {
-  if (!pokemon) return {};
-
-  return {
-    id: pokemon.id,
-    name: pokemon.name,
-    imageUri: pokemon.pokemon_v2_pokemonsprites[0].sprites["official-artwork"]
-      .front_default as string,
-    types: pokemon.pokemon_v2_pokemontypes
-      ?.map((item) => item.pokemon_v2_type?.name)
-      .filter((item) => item) as PokemonTypes[],
-    evolutions:
-      pokemon.pokemon_v2_pokemonspecy?.pokemon_v2_evolutionchain
-        ?.pokemon_v2_pokemonspecies,
-  };
-};
-
 const usePokemon = (name: string) => {
   const { data, error, loading } = useQuery(GET_POKEMON, {
     variables: {
@@ -55,7 +35,9 @@ const usePokemon = (name: string) => {
     },
   });
 
-  const pokemon = parsePokemon(data?.pokemon_v2_pokemon[0]);
+  const pokemon = data?.pokemon_v2_pokemon[0]
+    ? parsePokemon(data?.pokemon_v2_pokemon[0])
+    : undefined;
 
   return {
     loading,
